begin merge with paternity test
create_cram_crai_table() as it is doesn't handle files with extra periods in their names, which is common in some files such as 1000 genomes. I'll be trying an alternative method to query google cloud that effectively combines the existing function with one taken from my old "paternity test" notebook. this commit is not functional, it's just the first steps of that combining.