update and change first GWAS BP notebook (#66)
Two types of changes: (1) Rearranged some wording to make the purpose of the notebook clearly distinct from the setup section. (2) Updated the configuration guide to reflect the current UI of Terra, and indicated that these changes MIGHT need to be made rather than DEFINITELY because it appears the suggestions in this notebook are the default anyway.