add warning to GRM (see add'l notes)
When I let this run for about 12 hours, it seemed to got stuck. However it did output "2020-11-25 17:25:16 Hail: INFO: Wrote all 666 blocks of 2727214 x 1028 matrix with block size 4096." at some point, and it appears this GRM is used by the Genesis GWAS workflow which ran successfully on 1.4.1 (but not master). This means either it ran successfully and never exited the cell correctly, or it got stuck and the partial output was enough to run the workflow correctly (but we can't trust the results). More investigation is needed.