tell the user to hold their horses on immediately submitted stuff
This fires when the input is valid, but cromwell hasn't spun up yet. It is a distinct situation from the input being invalid, which is what the previous commit was about. We tell them to re-run the whole notebook in the valid-but-not-started case as if they just rerun the report cell, fiss won't be called again, so the report will not be updated.