Add Beth's GWAS Blood Pressure notebooks (#68)
* Add herzogifed GWAS notebooks

* Update .gitlab-ci.yml

* pfb fix

* Update and change first GWAS BP notebook (#66)

Two types of changes: (1) Rearranged some wording to make the purpose of the notebook clearly distinct from the setup section. (2) Updated the configuration guide to reflect the current UI of Terra, and indicated that these changes MIGHT need to be made rather than DEFINITELY because it appears the suggestions in this notebook are the default anyway.

* Update notebook config to match Terra's new setup

You may notice the workers memory change from 15 to 30. this is because 15 is no longer an option for spark clusters on terra; it uses a drop down menu now.

* Add warning to GRM

It can take more 12 hours after all!

* Remove outdated kernal info, added note to check you're running hail

* Fix bug in PCA no matter what lint thinks, added timing notes, small spacing fixes

* Revert table to HTML in order to left-align it

A left alignment allows users to view the table while setting environment options. Without this, it just isn't possible to see both at once unless you have a 4K monitor and are using Terra in full screen. This is bad because you must set the entire environment at once.

* leftalign first notebook's table, hotfix disk size to 500

* Make tests pass for GWAS_blood_pressure_p2 (#71)

* Make linting, mypy, and tests pass

* Mock tests

* Link fixes, add matplot to requirements

Co-authored-by: Brian Hannafious <32105697+xbrianh@users.noreply.github.com>