Add old functions, strips gs://, move table deletion
Previously, we would set WORKSPACE_BUCKET to something beginning with "gs://" and use that, with the gs:// included, throughout this notebook. And, currently, os.environ['WORKSPACE_BUCKET'] on Terra will return a URI with "gs://" at the front of it.
It seems that something might have changed with how Google accepts URIs as we now have to remove that gs:// part of the URI to use list_blobs(), or perhaps it was already getting removed in an older version but that was lost at some point.